Changed the way we handle NETSTACK shortcuts
Instead of conditionally replacing invocations of stack-hungry functions
we wrap the function itself inside a guard. When SHORTCUTS are defined,
we get rid of the function by defining it as something else across the
board. This has the positive side effect that the optimisation
propagates too all mac/rdc drivers.

This change also reduces our diff from master quite a bit